+++ spec/features/admin_checks_maths_and_physics_claim_spec.rb
@@ -3,13 +3,13 @@ require "rails_helper"
 RSpec.feature "Admin checking a Maths & Physics claim" do
   let(:user) { create(:dfe_signin_user) }
 
+  let!(:claim) { create(:claim, :submitted, policy: MathsAndPhysics) }
+
   before do
     sign_in_to_admin_with_role(DfeSignIn::User::SERVICE_OPERATOR_DFE_SIGN_IN_ROLE_CODE, user.dfe_sign_in_id)
   end
 
   scenario "service operator checks and approves a Maths & Physics claim" do
-    claim = create(:claim, :submitted, policy: MathsAndPhysics)
-
     visit admin_claims_path
     find("a[href='#{admin_claim_tasks_path(claim)}']").click
 
@@ -46,6 +46,33 @@ RSpec.feature "Admin checking a Maths & Physics claim" do
     expect(page).to have_content("Claim has been approved successfully")
     expect(claim.decision).to be_approved
     expect(claim.decision.created_by).to eq(user)
+  end
+
+  scenario "service operator can check a claim with matching details" do
+    claim_with_matching_details = create(:claim, :submitted,
+      teacher_reference_number: claim.teacher_reference_number,
+      policy: MathsAndPhysics)
+
+    visit admin_claims_path
+    find("a[href='#{admin_claim_tasks_path(claim)}']").click
+
+    expect(page).to have_content("1. Qualifications")
+    expect(page).to have_content("2. Employment")
+    expect(page).to have_content("3. Matching details")
+    expect(page).to have_content("4. Decision")
+
+    click_on I18n.t("maths_and_physics.admin.tasks.matching_details.summary")
+
+    expect(page).to have_content(I18n.t("maths_and_physics.admin.tasks.matching_details.question"))
+    expect(page).to have_content(claim_with_matching_details.reference)
+    expect(page).to have_content("Teacher reference number")
+
+    choose "Yes"
+    click_on "Save and continue"
+
+    expect(claim.tasks.find_by!(name: "matching_details").passed?).to eq(true)
+
+    expect(page).to have_content("Claim decision")
   end
 
   scenario "service operator sees an error if they don't choose a Yes/No option on a check" do